refactors the project architecture
