+++ controllers/auth/register.js
@@ -2,7 +2,8 @@ const bcryptjs = require("bcryptjs");
 const gravatar = require("gravatar");
 const { nanoid } = require("nanoid");
 
-const { HttpError, ctrlWrapper, sendEmail } = require("../../helpers");
+const { HttpError, ctrlWrapper } = require("../../helpers");
+const { sendEmail } = require("../../services/email");
 const { User } = require("../../models/user");
 
 const BASE_URL = process.env.BASE_URL;
