+++ helpers/index.js
@@ -1,11 +1,9 @@
 const HttpError = require("./HttpError");
 const ctrlWrapper = require("./ctrlWrapper");
 const handleMongooseError = require("./handleMongooseError");
-const sendEmail = require("./sendEmail");
 
 module.exports = {
   HttpError,
   ctrlWrapper,
   handleMongooseError,
-  sendEmail,
 };
