+++ services/email/index.js
@@ -1,0 +1,5 @@
+const sendEmail = require("./sendEmail");
+
+module.exports = {
+  sendEmail,
+};
